+++ src/main/java/use_case/execute_buy/ExecuteBuyInteractor.java
@@ -4,7 +4,7 @@ import entity.Stock;
 import entity.StockMarket;
 import entity.Transaction;
 import entity.User;
-import use_case.session.SessionService;
+import use_case.session.SessionManager;
 
 import java.util.Date;
 
@@ -23,16 +23,20 @@ public class ExecuteBuyInteractor implements ExecuteBuyInputBoundary {
 
     @Override
     public void execute(ExecuteBuyInputData data) {
-        User currentUser = dataAccess.getUser("user name");
-        String ticker = data.ticker();
-        int quantity = data.quantity();
-
+        // get username by credential
         try {
+            String username = SessionManager.Instance()
+                    .getUsername(data.credential())
+                    .orElseThrow(ValidationException::new);
+            User currentUser = dataAccess.getUser(username);
+            String ticker = data.ticker();
+            int quantity = data.quantity();
+
             Stock stock = StockMarket.Instance().getStock(ticker).orElseThrow(StockNotFoundException::new);
             if (currentUser.getBalance() >= stock.getPrice() * quantity) {
                 // then the transaction is valid
                 Date timestamp = new Date();
-                Transaction transaction = new Transaction(timestamp, ticker, quantity, stock.getPrice());
+                Transaction transaction = new Transaction(timestamp, ticker, quantity, stock.getPrice(), "buy");
                 // call portfolio to add the transaction
                 currentUser.getPortfolio().addTransaction(transaction);
                 // TODO: return some fancy output data
@@ -41,6 +45,8 @@ public class ExecuteBuyInteractor implements ExecuteBuyInputBoundary {
             } else {
                 throw new InsufficientFundsException();
             }
+        } catch (ValidationException e) {
+            outputPresenter.prepareValidationErrorView();
         } catch (InsufficientFundsException e) {
             outputPresenter.prepareInsufficientFundsView();
         } catch (StockNotFoundException e) {
@@ -52,4 +58,6 @@ public class ExecuteBuyInteractor implements ExecuteBuyInputBoundary {
     static class InsufficientFundsException extends Exception {}
 
     static class StockNotFoundException extends Exception {}
+
+    static class ValidationException extends Exception {}
 }